requirements.txt
albumentations==1.0.3
anyio==3.3.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==3.3.1
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
cycler==0.10.0
dataclasses==0.6
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
future==0.18.2
geffnet==1.0.2
greenstalk==2.0.0
h5py==2.10.0
idna==3.2
imageio==2.9.0
importlib-metadata==3.10.1
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.10.1
jupyterlab==3.1.1
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
munch==2.5.0
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.2
nose==1.3.7
notebook==6.4.0
numpy==1.21.1
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
packaging==21.0
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pretrainedmodels==0.7.4
prometheus-client==0.11.0
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.26.0
requests-unixsocket==0.2.0
resnest==0.0.5
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.4.1
Send2Trash==1.7.1
six==1.16.0
sniffio==1.2.0
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.2.0
tifffile==2021.7.30
timm==0.4.12
torch==1.7.0+cu101
torchaudio==0.7.0
torchvision==0.8.1+cu101
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
zipp==3.5.0
